move selfcare path in route const
